Move `bingcd` to `modular`
