+++ src/modular/bingcd.rs
