+++ src/modular/bingcd/extension.rs
